+++ example/lib/main.dart
@@ -31,6 +31,14 @@ class MyHomePage extends StatefulWidget {
 }
 
 class _MyHomePageState extends State<MyHomePage> {
+  SearchItem controllerStartWithValue = SearchItem(
+    items: [
+      Item(ModelExample(name: 'Tiago', age: 37), true),
+      Item(ModelExample(name: 'Mel', age: 3), false),
+      Item(ModelExample(name: 'Monique', age: 30), false),
+      Item(ModelExample(name: 'Timothy', age: 0), false),
+    ],
+  );
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -237,6 +245,44 @@ class _MyHomePageState extends State<MyHomePage> {
                   ),
                 );
               },
+            ),
+            SizedBox(
+              height: 50,
+            ),
+            buildInformation(information: 'Programmatically change the list of items'),
+            EasySearch(
+              onSearch: (text) {
+                print('Filter Query: $text');
+                return getData(name: text);
+              },
+              startWithValue: true,
+              searchResultSettings: SearchResultSettings(
+                padding: EdgeInsets.only(left: 8.0, top: 8.0, right: 8.0),
+                label: LabelSettings.searchLabel(value: 'People'),
+              ),
+              filterPageSettings: FilterPageSettings(
+                searchOnShow: false,
+              ),
+              controller: controllerStartWithValue,
+            ),
+            Center(
+              child: RaisedButton(
+                onPressed: () {
+                  //1ª Create the new listToFill
+                  List<Item<ModelExample>> listToFill = [
+                    Item(ModelExample(name: 'ABC 123', age: 3), true),
+                    Item(ModelExample(name: 'ACB 132', age: 13), false),
+                    Item(ModelExample(name: 'BAC 213', age: 23), false),
+                    Item(ModelExample(name: 'BCA 231', age: 33), false),
+                    Item(ModelExample(name: 'CAB 312', age: 43), false),
+                    Item(ModelExample(name: 'CBA 321', age: 53), false),
+                  ];
+
+                  //2ª Update controller with new listToFill
+                  controllerStartWithValue.changingValues(listToFill);
+                },
+                child: Text('Changing list'),
+              ),
             ),
             SizedBox(
               height: 50,